Run clippy in the CI

----
#### AI description  (iteration 1)
#### PR Classification
Code cleanup and CI enhancement

#### PR Summary
This pull request introduces code cleanup by replacing `write` with `write_all` for better error handling and adds Clippy linting to the CI pipeline.
- `prototype/transport/transport-lib/src/connection/transport/tls_transport.rs`: Replaced `write` with `write_all` for better error handling.
- `prototype/connect-sync/connect-sync-lib/src/transport.rs`: Replaced `write` with `write_all` for better error handling.
- `prototype/connect-sync/connect-sync-lib/src/connection.rs`: Replaced `write` with `write_all` for better error handling.
- `prototype/transport/transport-lib/src/connection.rs`: Replaced `write` with `write_all` for better error handling.
- `/.pipeline/RustPrototype.yml`: Added Clippy linting to the CI pipeline.

+++ prototype/connect-sync/connect-sync-lib/src/prelogin.rs
@@ -79,7 +79,7 @@ impl Encode<BytesMut> for PreloginMessage {
 
         // encryption
         fields.push((PRELOGIN_ENCRYPTION, 0x01)); // encryption
-        data_cursor.write_u8(self.encryption as u8)?;
+        data_cursor.write_u8(self.encryption)?;
 
         // threadid
         fields.push((PRELOGIN_THREADID, 0x04)); // thread id
